inner_notificationをnotificationに変更
モデルやDBがすべてinner_notificationだと分かりづらい
ので、Wrapperクラスの方をnotification_facadeに変更
