+++ app/models/check_callbacks.rb
@@ -3,7 +3,7 @@
 class CheckCallbacks
   def after_create(check)
     if check.sender != check.reciever
-      Notification.checked(check)
+      NotificationFacade.checked(check)
     end
   end
 end
