+++ db/migrate/20190816100249_rename_inner_notifications_to_notifications.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RenameInnerNotificationsToNotifications < ActiveRecord::Migration[5.2]
+  def change
+    rename_table :inner_notifications, :notifications
+  end
+end
