+++ test/models/notification_test.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class NotificationTest < ActiveSupport::TestCase
+  test "should be valid" do
+    assert notifications(:notification_commented).valid?
+    assert notifications(:notification_checked).valid?
+    assert notifications(:notification_mentioned).valid?
+    assert notifications(:notification_read).valid?
+  end
+end
